refactor: computation and requested-data lifted
Things relevant to the application itself more than any given component have been lifted and relocated. Models that are predominantly useful for certain reducers and logics have been moved as well.
