--- src/Layout/API/index.ts
@@ -1,4 +1,0 @@
-export * from "./ApiRequests";
-export * from "./RequestedDataActions";
-export * from "./RequestedDataReducer";
-export * from "./RequestedDataState";
