--- src/Layout/Interfaces/Models/County.ts
@@ -1,10 +1,0 @@
-﻿import { Result } from "./Result";
-
-export interface County {
-    countyId: number;
-    name: string;
-    seats: number;
-    countryId: number;
-    electionId: number;
-    results: Result[];
-}
