--- src/Layout/Interfaces/Models/ElectionType.ts
@@ -1,8 +1,0 @@
-import { Election } from "./Election";
-
-export interface ElectionType {
-    countryId: number;
-    electionTypeId: number;
-    elections: Election[];
-    internationalName: string;
-}
