--- src/Layout/Interfaces/Models/index.ts
@@ -1,4 +1,0 @@
-export * from "./County";
-export * from "./Election";
-export * from "./ElectionType";
-export * from "./Result";
