+++ src/Layout/Interfaces/Payloads/index.ts
@@ -1,2 +1,1 @@
-export * from "./ComputationPayload";
 export * from "./SettingsPayload";
