--- src/Layout/Interfaces/Results/DistributionResult.ts
@@ -1,9 +1,0 @@
-import { Dictionary } from "../Utilities";
-import { SeatResult } from "./SeatResult";
-
-export interface DistributionResult {
-    /** A dictionary taking partyCodes and returning the matching number of seats won in this distribution */
-    seatsWon: Dictionary<number>;
-    /** List of information regarding the distribution of the individual seats */
-    seatResults: SeatResult[];
-}
