--- src/Layout/Interfaces/Results/PartyRestQuotients.ts
@@ -1,6 +1,0 @@
-import { LevelingSeat } from "./LevelingSeat";
-
-export interface PartyRestQuotients {
-    partyCode: string;
-    levelingSeats: LevelingSeat[];
-}
