--- src/Layout/Interfaces/Results/Result.ts
@@ -1,11 +1,0 @@
-﻿export interface Result {
-    resultId: number;
-    votes: number;
-    percentage: number;
-    electionId: number;
-    partyId: number;
-    countyId: number;
-    countyName: string;
-    partyCode: string;
-    partyName: string;
-}
