--- src/Layout/Interfaces/Results/SeatPartyResult.ts
@@ -1,6 +1,0 @@
-export interface SeatPartyResult {
-    /** The partyCode of the participating party */
-    partyCode: string;
-    /** The quotient the party got in the distribution */
-    quotient: number;
-}
