--- src/Layout/Interfaces/Results/index.ts
@@ -1,9 +1,0 @@
-export * from "./DistributionResult";
-export * from "./DistrictResult";
-export * from "./LagueDhontResult";
-export * from "./LevelingSeat";
-export * from "./PartyRestQuotients";
-export * from "./PartyResult";
-export * from "./Result";
-export * from "./SeatPartyResult";
-export * from "./SeatResult";
