--- src/Layout/Interfaces/Utilities/Dictionary.ts
@@ -1,3 +1,0 @@
-﻿export interface Dictionary<T> {
-    [key: string]: T;
-}
