--- src/Layout/Interfaces/Utilities/index.ts
@@ -1,1 +1,0 @@
-export * from "./Dictionary";
