--- src/Layout/Logic/index.ts
@@ -1,4 +1,0 @@
-export * from "./AlgorithmUtils";
-export * from "./DictionaryUtilities";
-export * from "./DistributeLevelingSeats";
-export * from "./LagueDhont";
