+++ src/Layout/Presentation/Views/SingleDistrict.tsx
@@ -1,7 +1,7 @@
-import { DistrictResult, PartyResult } from "../../Interfaces/Results";
 import * as React from "react";
 import ReactTable from "react-table";
 import { toSum } from "../Utilities/ReduceUtilities";
+import { DistrictResult, PartyResult } from "../../../computation";
 
 export interface SingleDistrictProps {
     districtResults: DistrictResult[];
