+++ src/Layout/Types/ActionTypes.ts
@@ -1,16 +1,7 @@
-﻿export enum RequestedDataAction {
-    InitializeRequestedData = "INITIALIZE_REQUESTED_DATA"
-}
-
-export enum SettingAction {
+﻿export enum SettingAction {
     InitializeSettings = "INITIALIZE_SETTINGS",
     UpdateSettings = "UPDATE_SETTINGS",
     ToggleAutoCompute = "TOGGLE_AUTO_COMPUTE"
-}
-
-export enum ComputationAction {
-    InitializeComputation = "INITIALIZE_COMPUTATION",
-    UpdateResults = "UPDATE_CALCULATION"
 }
 
 export enum PresentationAction {
