+++ src/reducers/state.ts
@@ -1,7 +1,7 @@
 import { RouterState } from "react-router-redux";
-import { ComputationState } from "../Layout/Computation";
+import { ComputationState } from "../computation";
 import { SettingsState } from "../Layout/ComputationSettings";
-import { RequestedDataState } from "../Layout/API";
+import { RequestedDataState } from "../requested-data";
 import { PresentationState } from "../Layout/PresentationSettings";
 
 export interface RootState {
